refactor into new testing object
